下書き機能完成

+++ HEW/site/index.py
@@ -305,13 +305,6 @@ def SellPage():
 def SellConfirm():
     if request.method == 'POST':
         
-        if  request.args.get('sellid'):
-            print('あるあるあるあるあるあるあるあるあるある')
-            sellid = request.args.get('sellid')
-            return redirect(url_for('DelDraftImg',sellid=sellid))
-        else:
-            print('なしなしなしなしなしなしなしなしなしなし')
-        
         conn = conn_db()
         cursor = conn.cursor()
         
@@ -357,11 +350,15 @@ def SellConfirm():
         cursor.execute(ProfIMG_Select)
         icon = cursor.fetchone()[0]
         
+        
         # 画像
         input_imgs = request.files.getlist('uploadInput')
-        select = request.form['select']
         print('new:',input_imgs)
-        print('select',select)
+        if request.form['select']:
+            select = request.form['select']
+            print('select:',select)
+        else:
+            select = None
             
         # 商品名
         selltit = request.form['selltit']
@@ -385,6 +382,7 @@ def SellConfirm():
             
         # カテゴリー
         scategoryid = request.form['scategoryid']
+        print('category:',scategoryid)
         SCategory_Select = '''
         SELECT Name
         FROM Scategory
@@ -395,6 +393,7 @@ def SellConfirm():
             
         # サイズ
         postage = request.form['postage']
+        print('postage:',postage)
         Postage_Select = '''
         SELECT Size
         FROM Postage
@@ -430,6 +429,21 @@ def SellConfirm():
             imgs = None 
         # ========== 画像処理 ==========
             
+        
+        if request.form['select']:
+            print('下書きの画像処理')
+            sellid = None
+        else:
+            sellid = request.form['sellid']
+            SellIMG_Select = '''
+            SELECT SellIMG 
+            FROM SellIMG
+            WHERE SellID = {0};
+            '''.format(sellid)
+            cursor.execute(SellIMG_Select)
+            sellimgs = cursor.fetchall()
+            imgs = [img[0] for img in sellimgs]
+            
         sell_data = [selltit,overview,SCategoryName,PostageSize,StatusName,price]
         form_data = [imgs,selltit,overview,scategoryid,postage,status,price]
         
@@ -450,7 +464,7 @@ def SellConfirm():
         conn.close()
         return render_template('sell_confirm.html', icon=icon, UserName=UserName,select=select,
                 imgs=imgs, sell_data=sell_data, form_data=form_data, Address=Address, tags=tags,
-                style=style, layout_value=layout_value)
+                style=style, layout_value=layout_value, sellid=sellid)
 
 # /sell/
 @app.route('/sell/', methods=['POST'])
@@ -463,6 +477,15 @@ def Sell():
         you_list = session.get('you')
         if you_list:
             AccountID, UserName, MailAddress = you_list[0]
+        
+        if request.form['sellid']:
+            
+            draftid = request.form['sellid']
+            Sell_Update = '''
+            UPDATE sell SET draft = 1 WHERE SellID = {0};
+            '''.format(draftid)
+            cursor.execute(Sell_Update)
+            return redirect(url_for('IndexPage'))
         
         # ========== フォーム ==========
         # メイン画像
@@ -2041,27 +2064,28 @@ def DraftPage():
         style=style, layout_value=layout_value)
    
 # /deleteimg
-@app.route('/deleteimg', methods=['GET'])
+@app.route('/deleteimg', methods=['POST'])
 def DelDraftImg():
-    print('''
-          /deleteimg が実行
-          ''')
-    conn = conn_db()
-    cursor = conn.cursor()
-    
-    sellid = request.args.get('sellid')
-    
-    SellIMG_Delete = '''
-    DELETE FROM SellIMG
-    WHERE SellID = {0};
-    '''.format(sellid)
-    cursor.execute(SellIMG_Delete)
-    print('実行:',SellIMG_Delete)
-    
-    conn.commit()
-    cursor.close()
-    conn.close()
-    return redirect(url_for('SellPage', sellid=sellid))
+    if request.method == 'POST':
+        print('''
+            /deleteimg が実行
+            ''')
+        conn = conn_db()
+        cursor = conn.cursor()
+        
+        sellid = request.form['sellid']
+        
+        SellIMG_Delete = '''
+        DELETE FROM SellIMG
+        WHERE SellID = {0};
+        '''.format(sellid)
+        cursor.execute(SellIMG_Delete)
+        print('実行:',SellIMG_Delete)
+        
+        conn.commit()
+        cursor.close()
+        conn.close()
+        return redirect(url_for('SellPage', sellid=sellid))
         
 # /personal
 @app.route('/personal')
